+++ CHANGELOG.rst
@@ -3,6 +3,11 @@ Changelog
 
 .. currentmodule:: flask_restplus
 
+Current
+-------
+
+- Nothing yet
+
 0.11.0 (2018-05-16)
 -------------------